api/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pymongo==4.7.0
cassandra-driver==3.29.0
python-dotenv==1.0.0